made pr changes
